Adding a default, empty sofe configuration.
The background for this change is that when jspm is building code, it is running
in an environment where System.sofe might not be defined. The example of this is
a sofe service that is using `jspm build` to bundle itself. The service itself hasn't
set up sofe, but rather is just wanting to produce a distributable.